rename mainValueCard to SwipedCard; move SwipedCard to common components
